+++ lib/responders/bullmq-responders.ts
@@ -17,12 +17,13 @@ function paginate(
   start = start || 0;
   end = end || -1;
   return (<any>queue)[method](start, end, opts).then(function (jobs: Job[]) {
-    respond(ws, messageId, jobs);
+    respond(ws, Date.now(), messageId, jobs);
   });
 }
 
 async function respondJobCommand(ws: WebSocketClient, queue: Queue, msg: any) {
   const data = msg.data;
+  const startTime = Date.now();
   const job = await queue.getJob(data.jobId);
 
   switch (data.cmd) {
@@ -48,7 +49,7 @@ async function respondJobCommand(ws: WebSocketClient, queue: Queue, msg: any) {
         `Missing command ${data.cmd}. Too old version of taskforce-connector?`
       );
   }
-  respond(ws, msg.id);
+  respond(ws, startTime, msg.id);
 }
 
 async function respondQueueCommand(
@@ -57,14 +58,15 @@ async function respondQueueCommand(
   msg: any
 ) {
   const data = msg.data;
+  const startTime = Date.now();
   switch (data.cmd) {
     case "getJob":
       const job = await queue.getJob(data.jobId);
-      respond(ws, msg.id, job);
+      respond(ws, startTime, msg.id, job);
       break;
     case "getJobCounts":
       const jobCounts = await queue.getJobCounts();
-      respond(ws, msg.id, jobCounts);
+      respond(ws, startTime, msg.id, jobCounts);
       break;
     case "getMetrics":
       const metrics = await (<any>queue).getMetrics(
@@ -72,8 +74,14 @@ async function respondQueueCommand(
         data.start,
         data.end
       );
-      respond(ws, msg.id, metrics);
+      respond(ws, startTime, msg.id, metrics);
       break;
+    case "getDependencies":
+      const dependencies = await queue.getDependencies(data.parentId, data.type, data.start, data.end);
+      respond(ws, startTime, msg.id, dependencies);
+      break;
+
+    case "getWaitingChildren":
     case "getWaiting":
     case "getActive":
     case "getDelayed":
@@ -86,63 +94,67 @@ async function respondQueueCommand(
 
     case "getJobLogs":
       const logs = await queue.getJobLogs(data.jobId, data.start, data.end);
-      respond(ws, msg.id, logs);
+      respond(ws, startTime, msg.id, logs);
 
+    case "getWaitingChildrenCount":
     case "getWaitingCount":
     case "getActiveCount":
     case "getDelayedCount":
     case "getCompletedCount":
     case "getFailedCount":
     case "getRepeatableCount":
-    case "getWorkersCount":
       const count = await (<any>queue)[data.cmd]();
-      respond(ws, msg.id, count);
+      respond(ws, startTime, msg.id, count);
+      break;
+    case "getWorkersCount":
+      const workers = await queue.getWorkers();
+      respond(ws, startTime, msg.id, workers.length);
       break;
     case "removeRepeatableByKey":
       await queue.removeRepeatableByKey(data.key);
-      respond(ws, msg.id);
+      respond(ws, startTime, msg.id);
       break;
     case "add":
       const [name, jobData, opts] = data.args as [string, object, object];
       await queue.add(name, jobData, opts);
-      respond(ws, msg.id);
+      respond(ws, startTime, msg.id);
       break;
     case "empty":
       await queue.drain();
-      respond(ws, msg.id);
+      respond(ws, startTime, msg.id);
       break;
     case "pause":
       await queue.pause();
-      respond(ws, msg.id);
+      respond(ws, startTime, msg.id);
       break;
     case "resume":
       await queue.resume();
-      respond(ws, msg.id);
+      respond(ws, startTime, msg.id);
       break;
     case "isPaused":
       const isPaused = await queue.isPaused();
-      respond(ws, msg.id, isPaused);
+      respond(ws, startTime, msg.id, isPaused);
       break;
     case "obliterate":
       await queue.obliterate();
-      respond(ws, msg.id);
+      respond(ws, startTime, msg.id);
       break;
     case "clean":
       await queue.clean(data.grace, data.limit, data.status);
-      respond(ws, msg.id);
+      respond(ws, startTime, msg.id);
       break;
     case "retryJobs":
       await (<any>queue).retryJobs({
         status: data.status,
         count: data.count,
       });
-      respond(ws, msg.id);
+      respond(ws, startTime, msg.id);
       break;
     default:
       console.error(
         `Missing command ${data.cmd}. Too old version of taskforce-connector?`
       );
-      respond(ws, msg.id, null);
+      respond(ws, startTime, msg.id, null);
   }
 }