+++ lib/responders/respond.ts
@@ -1,9 +1,9 @@
 import { WebSocketClient } from "../ws-autoreconnect";
 
-export const respond = (ws: WebSocketClient, id: string, data: any = {}) => {
+export const respond = (ws: WebSocketClient, startTime: number, id: string, data: any = {}) => {
   const response = JSON.stringify({
     id,
     data,
   });
-  ws.send(response);
+  ws.send(response, startTime);
 };