+++ lib/version-checker.ts
@@ -1,0 +1,33 @@
+import { gt } from 'semver';
+import { yellow } from "chalk";
+
+/**
+ * Check if the current version of taskforce is the latest version
+ * @param name - package name
+ * @param version - current version of taskforce
+ */
+export const versionChecker = async (name: string, version: string) => {
+
+  const latestVersion = (await import('latest-version')).default;
+
+  try {
+    const newestVersion = await latestVersion(name);
+
+    if (gt(newestVersion, version)) {
+      console.error(
+        yellow(
+          "New version " +
+          newestVersion +
+          " of taskforce available, please upgrade with yarn global add taskforce-connector"
+        )
+      );
+    }
+  } catch (err) {
+    console.error(
+      yellow(
+        "Error checking for latest version of taskforce"
+      ),
+      err
+    );
+  }
+};