+++ lib/ws-autoreconnect.ts
@@ -9,7 +9,7 @@ export class WebSocketClient {
   private autoReconnectInterval = 5 * 1000; // ms
   private url: string;
   private opts: object;
-  private instance: any;
+  private instance: WebSocket;
   private pingTimeout: NodeJS.Timeout;
 
   open(url: string, opts: object) {
@@ -62,9 +62,25 @@ export class WebSocketClient {
     });
   }
 
-  send(data: string, option?: object) {
+  send(data: string, startTime: number, option?: {
+    mask?: boolean;
+    binary?: boolean;
+    compress?: boolean;
+    fin?: boolean;
+  }) {
     try {
-      this.instance.send(data, option);
+      this.instance.send(data, option, (err: Error) => {
+        if (err) {
+          console.log(
+            `${chalk.yellow("WebSocket:")} ${chalk.red("send error", err)}`
+          );
+        } else {
+          console.log(
+            `${chalk.yellow("WebSocket:")} ${chalk.blue("data sent successfully in")} ${chalk.green(
+              Date.now() - startTime)}ms`
+          );
+        }
+      });
     } catch (err) {
       this.instance.emit("error", err);
     }
@@ -74,7 +90,7 @@ export class WebSocketClient {
     var msg = err.message || "";
     console.log(
       chalk.yellow("WebSocket:") +
-        chalk.red(` ${msg} retry in ${this.autoReconnectInterval}ms`)
+      chalk.red(` ${msg} retry in ${this.autoReconnectInterval}ms`)
     );
     this.instance.removeAllListeners();
     setTimeout(() => {
@@ -95,13 +111,13 @@ export class WebSocketClient {
     }, HEARTBEAT_INTERVAL);
   }
 
-  onmessage = function(data: string, flags: object, num: number) {
+  onmessage = function (data: string, flags: object, num: number) {
     console.log("WebSocket: message", data, flags, num);
   };
-  onerror = function(e: Error) {
+  onerror = function (e: Error) {
     console.log("WebSocket: error", arguments);
   };
-  onclose = function(e: Error) {
+  onclose = function (e: Error) {
     console.log("WebSocket: closed", arguments);
   };
 }